rag-pdf-chatbot/requirements.txt

fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1

langchain>=0.2.3
langchain-community>=0.2.4
langchain-text-splitters>=0.0.1
chromadb>=0.5.0

sentence-transformers>=3.0.0
pypdf>=4.2.0
tqdm>=4.66.0
numpy>=1.26.0

openai>=1.38.0

# Optional retrieval extras (toggle via .env)
rank-bm25>=0.2.2
streamlit>=1.36.0
httpx>=0.27.0

# Dev
requests>=2.32.2
